Enable pip cache and build isolation
Disabling build isolation causes issues when different
projects have conflicting build requirements. The reason
for disabled build isolation was to build everything with
the pinned build dependencies. We achieve that with
PIP_CONSTRAINTS to pin the build dependencies.

We also re-enable the pip cache and use docker
mount type=cache for the pip cache. This improves
local builds and builds in a build farm where the cache
is reused across builds.